requirements.txt
annotated-types==0.7.0
anyio==4.8.0
apng==0.3.4
blinker==1.4
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.2
ExifRead==3.0.0
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
idna==3.6
importlib-metadata==4.6.4
intervaltree==3.1.0
jeepney==0.7.1
Jinja2==3.1.5
jiter==0.8.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
openai==1.59.7
openai-whisper==20240930
pillow==10.3.0
prettytable==3.10.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pypng==0.20220715.0
python-apt==2.4.0+ubuntu3
python-magic==0.4.27
python-tlsh==4.5.0
python-xmp-toolkit==2.0.2
pytz==2024.1
PyYAML==5.4.1
regex==2024.11.6
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
spark-parser==1.8.9
stegoveritas==1.10
stegoveritas-binwalk==2.1.2
stegoveritas-pfp==0.2.3
stegoveritas-py010parser==0.1.10
sympy==1.13.1
systemd-python==234
term-background==1.0.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uncompyle6==3.9.1
UNKNOWN @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
urllib3==2.2.1
wadllib==1.3.6
wcwidth==0.2.13
xdis==6.1.0
zipp==1.0.0
